+++ bdse-integration-tests/src/test/java/ru/csc/bdse/kv/PartitionedSingleFailureConsistentTest.java
@@ -40,11 +40,11 @@ public class PartitionedSingleFailureConsistentTest extends AbstractPartitionedK
 
     @Override
     protected float expectedKeysLossProportion() {
-        return (node0.getKeys("").size() + node1.getKeys("").size() + node2.getKeys("").size()) / 3;
+        return ((float)(node0.getKeys("").size() + node1.getKeys("").size() + node2.getKeys("").size())) / 3;
     }
 
     @Override
     protected float expectedUndeletedKeysProportion() {
-        return (node0.getKeys("").size() + node1.getKeys("").size() + node2.getKeys("").size()) / 3;
+        return ((float)(node0.getKeys("").size() + node1.getKeys("").size() + node2.getKeys("").size())) / 3;
     }
 }
